feat(core)!: map whole number protocol fields to int and long
The protocol has one numeric type because JSON has one, so every Number
became a double and callers read scene item ids, frame counts and byte
counts as floating point. Which fields are integral is not recoverable
from the definition: sceneItemId and inputVolumeMul are both Number with
a >= 0 restriction. The classification is therefore an explicit table,
not a rule over field names, because guessing wrong on a volume field
truncates it silently while an unlisted field only stays double.

32 field names become int or long, 10 stay double. A refresh that adds
an unlisted Number field reports OBSWSGEN012 rather than drifting in.

Verified against OBS 32.2.2 on both transports: MessagePack decodes the
retyped fields, which was the risk.

+++ ObsWebSocket.Tests/ObsWebSocketClientIntegrationTests.cs
@@ -445,7 +445,7 @@ public class ObsWebSocketClientIntegrationTests
             idResponse,
             $"Could not get SceneItemId for '{s_testOptions.TestInputName}' in scene '{s_testOptions.TestSceneName}'. Ensure it exists."
         );
-        double sceneItemId = idResponse.SceneItemId;
+        int sceneItemId = idResponse.SceneItemId;
 
         // Get the transform
         GetSceneItemTransformResponseData? transformResponse =
